Building initial infrastructure for QA Regression suite for spark-tk
Builds a set of files and scripts for installation of datasets, automatic
installation of spark-tk from the build server (not from source), execution
of the test cases, and a sample test case of confusion matrix to prove
the system is operational.

Also includes a set of development scripts in order to facilitate QA
test writing.