Converts the unittests and integration to work the same as integration tests (#221)
Switches unittests to use the py test test runner, export the coverage report to
the top level to be reaped in a standard way. The integration tests are converted
to directly use pytest, also adds code coverage.

All the coverage tests export their coverage information, this is used for
composite coverage.